.config/restud/pyproject.toml
[tool.poetry]
name = "restud"
version = "0.1.0"
description = ""
authors = ["Miklós Koren <koren@localhost>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
PyYAML = "^6.0.1"
pydantic = "^2.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
